requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
aiogram==2.25.1
python-dateutil~=2.8.2
SQLAlchemy~=2.0.25
cryptography~=41.0.7
aiosqlite~=0.19.0